requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
Flask==3.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-telegram-bot==20.3
sniffio==1.3.1
tornado==6.5.2
typing_extensions==4.15.0
Werkzeug==3.1.3
